requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pymongo==4.7.2
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
xgboost==2.0.3
torch==2.3.0
torchvision==0.18.0
tensorflow==2.16.1
shap==0.45.0
matplotlib==3.9.0
seaborn==0.13.2
jupyter==1.0.0
python-dotenv==1.0.1
pillow==10.3.0
python-multipart==0.0.9